+++ FE/src/app/component/stack/header/header.component.tsx
@@ -74,11 +74,11 @@ const HeaderComponent: FunctionComponent<HeaderComponentProps> = () => {
     return (
         <div className='mx-auto flex items-center justify-between sm:w-[1440px]'>
             <div css={cssMenu} className='space-x-8'>
-                <div><Link to={'/'}>TRANG CHỦ</Link></div>
-                <div><Link to={'/products'}>SẢN PHẨM</Link></div>
-                <div><Link to={'/'}>GIỚI THIỆU</Link></div>
-                <div><Link to={'/'}>LIÊN HỆ</Link></div>
-                <div><Link to={'/'}>VỀ CHÚNG TÔI</Link></div>
+                <div><Link className='hover:text-red-500' to={'/'}>TRANG CHỦ</Link></div>
+                <div><Link className='hover:text-red-500' to={'/'}>SẢN PHẨM</Link></div>
+                <div><Link className='hover:text-red-500' to={'/'}>GIỚI THIỆU</Link></div>
+                <div><Link className='hover:text-red-500' to={'/'}>LIÊN HỆ</Link></div>
+                <div><Link className='hover:text-red-500' to={'/'}>VỀ CHÚNG TÔI</Link></div>
 
 
             </div>
